pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-harness"
version = "0.1.0"
description = "Governance, retrieval, and enforcement harness for OpenClaw multi-agent memory"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
brain-validate-qmd = "plugins.qmd_guard.cli:main"
brain-validate-tags = "plugins.tag_guard.cli:main"
brain-validate-provenance = "plugins.provenance_guard.cli:main"
brain-retrieve = "plugins.retrieval_router.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
